runtime/requirements.txt
strands-agents[otel]
bedrock-agentcore
boto3
requests
aws-opentelemetry-distro
